directory name changed from header to navbar
