+++ frontend/client/src/components/navbar/components/Collections.tsx
