+++ frontend/client/src/components/navbar/components/Eyes.tsx
