+++ frontend/client/src/components/navbar/components/Lips.tsx
