+++ frontend/client/src/components/navbar/components/SearchInput.tsx
