+++ frontend/client/src/components/navbar/components/Skin.tsx
