+++ frontend/client/src/components/navbar/components/skin/Skin.tsx
@@ -1,0 +1,9 @@
+import React from 'react'
+
+const Skin = () => {
+  return (
+    <div>Skin</div>
+  )
+}
+
+export default Skin
