+++ frontend/client/src/pages/main/Main.tsx
@@ -1,6 +1,6 @@
 import { Outlet } from "react-router-dom";
 
-import Navbar from "../../components/header/Navbar";
+import Navbar from "../../components/navbar/Navbar";
 import Footer from "../../components/footer/Footer";
 import ErrorBoundary from "../error/ErrorBoundary";
